+++ __tests___/Usuario.spec.ts
@@ -26,7 +26,6 @@ describe("Testes da classe model: Usuário", () => {
             expect(wrapper).toThrow(Error);
         })
     })
-
     describe("Função que retorna o nome do Usuário", () =>{
         it("Deve retornar o nome Erik quando chamado a função", () => {
             jest.spyOn(HashHelper, "hash").mockReturnValue("algum_hash");
@@ -34,9 +33,8 @@ describe("Testes da classe model: Usuário", () => {
             const usuario = new Usuario("Erik", "erik@email.com", "123456");
             
             expect(usuario.getNome()).toBe("Erik");
-        });
-    });
-
+        })
+    })
     describe("Função que retorna o email de um Usuário", () => {
         it("Deve retornar o email erik@email.com quando chamado a função", () => {
             jest.spyOn(HashHelper, "hash").mockReturnValue("algum_hash");
@@ -44,9 +42,8 @@ describe("Testes da classe model: Usuário", () => {
             const usuario = new Usuario("Erik", "erik@email.com", "123456");
             
             expect(usuario.getEmail()).toBe("erik@email.com");
-        });
-    });
-    
+        })
+    })
     describe("Função que retorna a senha de um Usuário", () =>{
         it("Deve retornar a senha de um Usuário com hash", () => {
             jest.spyOn(HashHelper, "hash").mockReturnValue("algum_hash");
@@ -54,25 +51,27 @@ describe("Testes da classe model: Usuário", () => {
             const usuario = new Usuario("Erik", "erik@email.com", "123456");
             
             expect(usuario.getSenha()).toEqual("algum_hash");
-        });
-    });
-
-    // describe("Função que altera o nome de um Usuário", () =>{
-    //     it("Deve alterar o nome do usuário para Erin quando chamado a função", () => {
-    //         const usuario = new Usuario("Erik", "erik@email.com", "123456");
-    //         usuario.setNome("Erin");
+        })
+    })
+    describe("Função que altera o nome de um Usuário", () =>{
+        it("Deve alterar o nome do usuário para Erin quando chamado a função", () => {
+            jest.spyOn(HashHelper, "hash").mockReturnValueOnce("algum_hash");
+            jest.spyOn(EmailHelper, "validarEmail").mockReturnValue(true);
+            const usuario = new Usuario("Erik", "erik@email.com", "123456");
+            usuario.setNome("Erin");
             
-    //         expect(usuario.getNome()).toBe("Erin");
-    //     });
-    // });
-    
-    // describe("Função que altera a senha de um Usuário", () =>{
-    //     it("Deve alterar a senha do usuário para 654321", () => {
-    //         const usuario = new Usuario("Erik", "erik@email.com", "123456");
-    //         usuario.setSenha("654321");
-    //         const hashSenha = HashSenha.hash("654321");
+            expect(usuario.getNome()).toBe("Erin");
+        })
+    })
+    describe("Função que altera a senha de um Usuário", () =>{
+        it("Deve alterar a senha do usuário para 654321", () => {
+            jest.spyOn(HashHelper, "hash").mockReturnValueOnce("algum_hash");
+            jest.spyOn(EmailHelper, "validarEmail").mockReturnValue(true);
+            const usuario = new Usuario("Erik", "erik@email.com", "123456");
+            jest.spyOn(HashHelper, "hash").mockReturnValueOnce("outro_hash");
+            usuario.setSenha("654321");
             
-    //         expect(usuario.getSenha()).toBe(hashSenha);
-    //     });
-    // });
+            expect(usuario.getSenha()).toBe("outro_hash");
+        })
+    })
 })